requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
ccxt==4.4.88
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dateparser==1.2.1
fastapi==0.115.12
frozenlist==1.6.2
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.4.4
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.0
